+++ services/src/test/java/org/jboss/windup/web/services/data/ServiceConstants.java
@@ -1,12 +1,10 @@
 package org.jboss.windup.web.services.data;
 
-import org.jboss.windup.web.services.rest.RegisteredApplicationEndpoint;
-
 /**
  * @author <a href="mailto:jesse.sightler@gmail.com">Jesse Sightler</a>
  */
-public abstract class ServiceConstants {
+public abstract class ServiceConstants
+{
     public static final String REST_BASE = "rest";
     public static final String FURNACE_REST_BASE = "rest-furnace";
-    public static final String REGISTERED_APP_ENDPOINT = RegisteredApplicationEndpoint.REGISTERED_APPLICATIONS;
 }
